updating for AGU 2017
Benefit = sum of next 5 years benefit ( 5 year project).
communities assume retreat rate of last year.
trying to make it an annual decision, but didn't work yet. - this is because it's too beneficial, so in 100 years with largest forcing, a community will nourish twice a year. not realistic.
SL rise increasing with each model run.